Correct project structure for gradle
Move src and build.gradle under project name `orion-sdk-java`.

This causes the name of the jar file to be "orion-sdk-java-<ver>.jar".

Set version to "0.1.0".

Correct absolute path of test resources to relative path.

Correct compilation errors.

Gradle command './gradlew jar',  './gradlew assemble', './gradlew clean'  work fine.

Gradle command './gradlew test' works only if the test server is started manually.
