+++ .gitignore
@@ -20,7 +20,7 @@ hs_err_pid*
 
 ### Gradle ###
 .gradle
-/build/
+orion-sdk-java/build/
 
 # Ignore Gradle GUI config
 gradle-app.setting
@@ -34,5 +34,13 @@ gradle-app.setting
 # # Work around https://youtrack.jetbrains.com/issue/IDEA-116898
 # gradle/wrapper/gradle-wrapper.properties
 #
+
+### Eclipse IDE ###
+.project
+.settings/
+orion-sdk-java/.classpath
+orion-sdk-java/.project
+orion-sdk-java/.settings/
+orion-sdk-java/bin/
 
 # End of https://www.gitignore.io/api/java,gradle
