+++ orion-sdk-java/build.gradle
@@ -43,3 +43,5 @@ tasks.named('test') {
     // Use JUnit Platform for unit tests.
     useJUnitPlatform()
 }
+
+version = '0.1.0'
