+++ orion-sdk-java/src/main/java/org/hyperledger/orion/sdk/exception/BaseException.java
