+++ orion-sdk-java/src/main/java/org/hyperledger/orion/sdk/security/certgen/TLSCertificateKeyPair.java
@@ -13,7 +13,7 @@
  *  limitations under the License.
  *
  */
-package org.hyperledger.fabric.sdk.security.certgen;
+package org.hyperledger.orion.sdk.security.certgen;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
