+++ orion-sdk-java/src/main/java/types/Query.java
