+++ orion-sdk-java/src/test/java/org/hyperledger/orion/sdk/LibraryTest.java
