+++ orion-sdk-java/src/test/java/org/hyperledger/orion/sdk/config/Replica.java
