--- src/test/resources/admin.key
@@ -1,5 +1,0 @@
------BEGIN EC PRIVATE KEY-----
-MHcCAQEEIFYjQF5mSL+oRcOdGPxULl44mbPxETiPWt/hY/cHFHuvoAoGCCqGSM49
-AwEHoUQDQgAE7BRSToz4fudVAIFP2sFSkCgC+9AXzuIAVJxmV5deHIyXBsN1Jy45
-ZHcEVSjU1waQVk4ntSRUZVAW0SyuptBG7A==
------END EC PRIVATE KEY-----
